Moved processTargets/targetUtil to processTargets/targets/util.
