+++ packages/cursorless-engine/src/processTargets/targets/InteriorTarget.ts
@@ -1,7 +1,7 @@
 import { Range } from "@cursorless/common";
 import { BaseTarget, MinimumTargetParameters } from "./BaseTarget";
 import { shrinkRangeToFitContent } from "../../util/selectionUtils";
-import { createContinuousRangeFromRanges } from "../targetUtil/createContinuousRange";
+import { createContinuousRangeFromRanges } from "./util/createContinuousRange";
 
 export interface InteriorTargetParameters extends MinimumTargetParameters {
   readonly fullInteriorRange: Range;
