+++ packages/cursorless-engine/src/processTargets/targets/SurroundingPairTarget.ts
@@ -14,7 +14,7 @@ import {
   getTokenLeadingDelimiterTarget,
   getTokenRemovalRange,
   getTokenTrailingDelimiterTarget,
-} from "../targetUtil/insertionRemovalBehaviors/TokenInsertionRemovalBehavior";
+} from "./util/insertionRemovalBehaviors/TokenInsertionRemovalBehavior";
 
 interface SurroundingPairTargetParameters extends CommonTargetParameters {
   /**
