+++ packages/cursorless-engine/src/processTargets/targets/util/createContinuousRange.ts
@@ -1,5 +1,5 @@
 import { Position, Range } from "@cursorless/common";
-import type { Target } from "../../typings/target.types";
+import type { Target } from "../../../typings/target.types";
 
 export function createContinuousRange(
   startTarget: Target,
