+++ packages/cursorless-engine/src/processTargets/targets/util/insertionRemovalBehaviors/TokenInsertionRemovalBehavior.ts
@@ -1,7 +1,7 @@
 import { Range, TextEditor } from "@cursorless/common";
-import type { Target } from "../../../typings/target.types";
-import { expandToFullLine } from "../../../util/rangeUtils";
-import { PlainTarget } from "../../targets";
+import type { Target } from "../../../../typings/target.types";
+import { expandToFullLine } from "../../../../util/rangeUtils";
+import { PlainTarget } from "../../PlainTarget";
 
 const leadingDelimiters = ['"', "'", "(", "[", "{", "<"];
 const trailingDelimiters = ['"', "'", ")", "]", "}", ">", ",", ";", ":"];
